+++ api/client/builder/bid.go
@@ -7,6 +7,7 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
+	v1 "github.com/prysmaticlabs/prysm/v5/proto/engine/v1"
 	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/runtime/version"
 )
@@ -29,6 +30,12 @@ type Bid interface {
 	IsNil() bool
 	HashTreeRoot() ([32]byte, error)
 	HashTreeRootWith(hh *ssz.Hasher) error
+}
+
+// BidElectra is an interface that exposes the newly added execution requests on top of the builder bid
+type BidElectra interface {
+	Bid
+	ExecutionRequests() (*v1.ExecutionRequests, error)
 }
 
 type signedBuilderBid struct {
@@ -288,5 +295,103 @@ func (b signedBuilderBidDeneb) Version() int {
 
 // IsNil --
 func (b signedBuilderBidDeneb) IsNil() bool {
+	return b.p == nil
+}
+
+type builderBidElectra struct {
+	p *ethpb.BuilderBidElectra
+}
+
+// WrappedBuilderBidElectra is a constructor which wraps a protobuf bid into an interface.
+func WrappedBuilderBidElectra(p *ethpb.BuilderBidElectra) (Bid, error) {
+	w := builderBidElectra{p: p}
+	if w.IsNil() {
+		return nil, consensus_types.ErrNilObjectWrapped
+	}
+	return w, nil
+}
+
+// Version --
+func (b builderBidElectra) Version() int {
+	return version.Electra
+}
+
+// Value --
+func (b builderBidElectra) Value() primitives.Wei {
+	return primitives.LittleEndianBytesToWei(b.p.Value)
+}
+
+// Pubkey --
+func (b builderBidElectra) Pubkey() []byte {
+	return b.p.Pubkey
+}
+
+// IsNil --
+func (b builderBidElectra) IsNil() bool {
+	return b.p == nil
+}
+
+// HashTreeRoot --
+func (b builderBidElectra) HashTreeRoot() ([32]byte, error) {
+	return b.p.HashTreeRoot()
+}
+
+// HashTreeRootWith --
+func (b builderBidElectra) HashTreeRootWith(hh *ssz.Hasher) error {
+	return b.p.HashTreeRootWith(hh)
+}
+
+// Header --
+func (b builderBidElectra) Header() (interfaces.ExecutionData, error) {
+	// We have to convert big endian to little endian because the value is coming from the execution layer.
+	return blocks.WrappedExecutionPayloadHeaderElectra(b.p.Header)
+}
+
+// ExecutionRequests --
+func (b builderBidElectra) ExecutionRequests() (*v1.ExecutionRequests, error) {
+	if b.p == nil {
+		return nil, consensus_types.ErrNilObjectWrapped
+	}
+	if b.p.ExecutionRequests == nil {
+		return nil, errors.New("ExecutionRequests is nil")
+	}
+	return b.p.ExecutionRequests, nil
+}
+
+// BlobKzgCommitments --
+func (b builderBidElectra) BlobKzgCommitments() ([][]byte, error) {
+	return b.p.BlobKzgCommitments, nil
+}
+
+type signedBuilderBidElectra struct {
+	p *ethpb.SignedBuilderBidElectra
+}
+
+// WrappedSignedBuilderBidElectra is a constructor which wraps a protobuf signed bit into an interface.
+func WrappedSignedBuilderBidElectra(p *ethpb.SignedBuilderBidElectra) (SignedBid, error) {
+	w := signedBuilderBidElectra{p: p}
+	if w.IsNil() {
+		return nil, consensus_types.ErrNilObjectWrapped
+	}
+	return w, nil
+}
+
+// Message --
+func (b signedBuilderBidElectra) Message() (Bid, error) {
+	return WrappedBuilderBidElectra(b.p.Message)
+}
+
+// Signature --
+func (b signedBuilderBidElectra) Signature() []byte {
+	return b.p.Signature
+}
+
+// Version --
+func (b signedBuilderBidElectra) Version() int {
+	return version.Electra
+}
+
+// IsNil --
+func (b signedBuilderBidElectra) IsNil() bool {
 	return b.p == nil
 }